Calculaton Screen modification, some main screen changes as well, some relevant database changes also

+++ app/src/main/java/com/energykhata/roomdb/dao/MeterDao.kt
@@ -18,5 +18,5 @@ interface MeterDao {
     suspend fun getMeters(): List<Meter>
 
     @Query("SELECT * FROM meters WHERE meter_id = :id")
-    suspend fun getMeter(id : Int): List<Meter>
+    suspend fun getMeter(id : Long): List<Meter>
 }
