+++ app/src/main/java/com/energykhata/ui/screens/history/HistoryScreen.kt
@@ -49,7 +49,7 @@ import com.energykhata.viewmodels.ReadingViewModel
 @Composable
 fun HistoryScreen(
     navController: NavHostController,
-    meterId: Int?,
+    meterId: Long?,
     meterRepository: MeterRepository,
     readingRepository: ReadingRepository,
 ) {
